common: attrs: combine attrs16 and attrs32
common: identity: removed

identity is no longer used. fabd just runs as the user that ran fab.
fab build && sudo fab install - this usage pattern will not be
supported.

common: snarf, stresc

+++ common/spinlock/spinlock.h
@@ -26,7 +26,7 @@
 
 static inline void spinlock_engage_as(int * v, int as)
 {
-  int x;
+  uint16_t x;
 
   if(*v != as)
   {
